[deploy] retrieving stored docker machine credentials

+++ deploy/deploy/secret/docker_machine.py
@@ -2,6 +2,7 @@ import os
 import shutil
 import subprocess32 as subprocess
 
+from deploy.secret.exceptions import SecretNotFoundError
 from deploy.secret.secret_dir import secret_dir_path
 from deploy.utils.utils import parse_shell_for_exports
 
@@ -21,4 +22,22 @@ def store_credentials(machine_name):
     for f in ['ca.pem', 'cert.pem', 'key.pem']:
         src = os.path.join(cert_path, f)
         dst = os.path.join(dir_path, f)
+        shutil.copy2(src, dst)
+
+
+def retrieve_credentials(machine_name, target_directory):
+    """Given a docker machine, retrieve the credentials
+    stored and copy it to the target directory.
+
+    @type machine_name: str
+    @type target_directory: str
+
+    @raises: SecretNotFoundError
+    """
+    dir_path = os.path.join(secret_dir_path(), 'docker_machines', machine_name)
+    if not os.path.exists(dir_path):
+        raise SecretNotFoundError
+    for f in ['ca.pem', 'cert.pem', 'key.pem']:
+        src = os.path.join(dir_path, f)
+        dst = os.path.join(target_directory, f)
         shutil.copy2(src, dst)
